Fixed runtime error by removing FS from PostList
Apparently the static site generation doesn't remove fully create
the HMTL for the post list component. For now I'm just manually
making the list that drives the PostList component. It's annoying to
have to do it manually but it's relativly simple.

There should be some way to list possible routes in Nextjs and show
a subset of them programatically. I probably just haven't found
what I need in the documentation yet.